Initial Feature Concept
This is the initial passing feature concept I hope to keep in as
an example and regression feature test.  The goal here is to be
able to be to include a custom created module from a schema that
allows your class to act like a model.

```ruby
class Person
  include GarbageFactory.model(YAML.read_file('person.yml'))
  # ...
end
```

The `Person` class should now have the properties defined in the
json-schema file.  I realize I am using yaml files to describe
the schemas; however, I don't think that's a bad thing ;)