First Pass Into Building an AST
This is the first part of what I have been thinking about doing to
create an AST that can be walked and used to pull in other source
files as needed.  The idea behind it is that the AST will be
initialized with the raw data and the source that was used to build
it.  One of the responsibilities of the source is to determine the
source format and scheme of where to pull in the initial schema. I
plan to use this to allow other referenced schemas to be pulled in
via the same scheme if they have a "relative" URI.